+++ src/x/hmm.learn.c
@@ -24,6 +24,7 @@ int main(int argc, char *argv[]) {
     int max_iterations = -1;
 
     int seed = -1;
+    int use_par = 1;
 
     if (argc < 2) {
         printf("\
@@ -44,6 +45,7 @@ HMM training\n\
                     2: cascade-2; random B\n\
                     3: cascade-3; random B\n\
     -s <val>      Seed for random numbers. Negative means random seed (%d, by default).\n\
+    -S            Use serialized impl (parallel impl, by default).\n\
     <sequence>... training sequences (max: %d)\n\
 \n\n",
                N, val_auto, hmm_epsilon, model_type, seed, MAX_SEQS
@@ -53,7 +55,7 @@ HMM training\n\
 
     int opc;
 
-    while (EOF != (opc = getopt(argc, argv, "N:t:e:a:I:s:"))) {
+    while (EOF != (opc = getopt(argc, argv, "N:t:e:a:I:s:S"))) {
         switch (opc) {
             case 'I':
                 if (sscanf(optarg, "%u", &max_iterations) == 0) {
@@ -92,6 +94,9 @@ HMM training\n\
                     return 1;
                 }
                 break;
+            case 'S':
+                use_par = 0;
+                break;
             case '?':
                 return 0;
         }
@@ -123,6 +128,7 @@ HMM training\n\
         hmm_epsilon,
         val_auto,
         max_iterations,
+        use_par,
         0  // hmm_learn_callback_t
     );
 }